Add float(Float)/double(Double) types to distinguish single vs double precision
Rename Expr.float(Double) to Expr.double(Double) and add Expr.float(Float)
backed by Swift's 32-bit Float, matching Clojure's type model. The float
function now coerces to single precision and double? returns false for
those values, while float? accepts both. All arithmetic promotes Float to
Double. Adds native double? predicate and removes the core.clj alias.

+++ Sources/SwishKit/Reader/Parser/Parser.swift
@@ -173,7 +173,7 @@ public class Parser {
             value = v
         }
         try advance()
-        return .float(value)
+        return .double(value)
     }
 
     private func parseRatio() throws -> Expr {
